+++ tests/test_obstacles_thinner_collocated.py
@@ -1,3 +1,4 @@
+#!/usr/bin/python
 """
 #                                                       o ... scalars
 #                          (n)                          - ... u velocities
@@ -25,8 +26,7 @@
 # +---> x                  (s)
 #
 """
-
-#!/usr/bin/python
+from __future__ import division
 
 # Standard Python modules
 from pyns.standard import *
@@ -129,9 +129,9 @@ def main(show_plot=True, time_steps=1800, plot_freq=180):
     block[3].kp = ny                 # k plus
 
     for o in range(0, 4):
-        for i in range(floor(block[o].im), floor(block[o].ip)):
-            for j in range(floor(block[o].jm), floor(block[o].jp)):
-                for k in range(floor(block[o].km), floor(block[o].kp)):
+        for i in range(int(floor(block[o].im)), int(floor(block[o].ip))):
+            for j in range(int(floor(block[o].jm)), int(floor(block[o].jp))):
+                for k in range(int(floor(block[o].km)), int(floor(block[o].kp))):
                     plates[i, j, k] = 1
 
 # =============================================================================
